+++ src/nodes/ListLiteral.ts
@@ -102,15 +102,17 @@ export default class ListLiteral extends Expression {
     }
 
     getItemType(context: Context): Type | undefined {
-        const expressions = this.values.filter(
-            (e) => e instanceof Expression,
-        ) as Expression[];
-        return expressions.length === 0
+        const types = this.values
+            .map((e) => {
+                if (e instanceof Spread) {
+                    const type = e.list?.getType(context);
+                    return type instanceof ListType ? type.type : undefined;
+                } else return e.getType(context);
+            })
+            .filter((type): type is Type => type !== undefined);
+        return types.length === 0
             ? undefined
-            : UnionType.getPossibleUnion(
-                  context,
-                  expressions.map((v) => v.getType(context)),
-              );
+            : UnionType.getPossibleUnion(context, types);
     }
 
     computeType(context: Context): Type {